refactor(winner): separate winner deciding methods to winner class
